+++ portal-web/test/com/liferay/portalweb/permissions/organizations/organization/vieworganization/orgrole/RemoveOrgRoleOrganizationsManageUsersTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>DefineOrgRoleOrganizationsManageUsersTest</title>
+<title>RemoveOrgRoleOrganizationsManageUsersTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">DefineOrgRoleOrganizationsManageUsersTest</td></tr>
+<tr><td rowspan="1" colspan="3">RemoveOrgRoleOrganizationsManageUsersTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
