+++ portal-web/test/com/liferay/portalweb/portal/PermissionsOrganizationsTestSuite.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.portalweb.permissionscp.organizations.organization;
+package com.liferay.portalweb.portal;
 
-import com.liferay.portalweb.permissionscp.organizations.organization.vieworganization.ViewOrganizationTestPlan;
-import com.liferay.portalweb.portal.BaseTestSuite;
+import com.liferay.portalweb.permissions.organizations.OrganizationsTestPlan;
+import com.liferay.portalweb.portal.login.LoginTests;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
@@ -23,12 +23,15 @@ import junit.framework.TestSuite;
 /**
  * @author Brian Wing Shun Chan
  */
-public class OrganizationTestPlan extends BaseTestSuite {
+public class PermissionsOrganizationsTestSuite extends BaseTestSuite {
 
 	public static Test suite() {
 		TestSuite testSuite = new TestSuite();
 
-		testSuite.addTest(ViewOrganizationTestPlan.suite());
+		testSuite.addTest(LoginTests.suite());
+		testSuite.addTest(OrganizationsTestPlan.suite());
+
+		testSuite.addTestSuite(StopSeleniumTest.class);
 
 		return testSuite;
 	}
