+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/organizations/organization/addorganization/TearDownOrganizationTest.html
@@ -104,11 +104,6 @@
 	<td>Update Associations</td>
 </tr>
 <tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
 	<td>label</td>
 	<td>Users1NotFound</td>
 	<td></td>
@@ -142,11 +137,6 @@
 	<td>assertConfirmation</td>
 	<td>Are you sure you want to delete this? It will be deleted immediately.</td>
 	<td></td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
@@ -219,11 +209,6 @@
 	<td>Update Associations</td>
 </tr>
 <tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
 	<td>label</td>
 	<td>Users2NotFound</td>
 	<td></td>
@@ -257,11 +242,6 @@
 	<td>assertConfirmation</td>
 	<td>Are you sure you want to delete this? It will be deleted immediately.</td>
 	<td></td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
@@ -334,11 +314,6 @@
 	<td>Update Associations</td>
 </tr>
 <tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
 	<td>label</td>
 	<td>Users3NotFound</td>
 	<td></td>
@@ -372,11 +347,6 @@
 	<td>assertConfirmation</td>
 	<td>Are you sure you want to delete this? It will be deleted immediately.</td>
 	<td></td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
@@ -449,11 +419,6 @@
 	<td>Update Associations</td>
 </tr>
 <tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
 	<td>label</td>
 	<td>Users4NotFound</td>
 	<td></td>
@@ -487,11 +452,6 @@
 	<td>assertConfirmation</td>
 	<td>Are you sure you want to delete this? It will be deleted immediately.</td>
 	<td></td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
@@ -564,11 +524,6 @@
 	<td>Update Associations</td>
 </tr>
 <tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
 	<td>label</td>
 	<td>Users5NotFound</td>
 	<td></td>
@@ -602,11 +557,6 @@
 	<td>assertConfirmation</td>
 	<td>Are you sure you want to delete this? It will be deleted immediately.</td>
 	<td></td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
 </tr>
 <tr>
 	<td>label</td>
